refactor(openclaw): split dae into control-plane modules
